requirements.txt
# Core RAG System Dependencies for macOS
# Tested with Python 3.14 on macOS

# Vector Database and Embeddings
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0

# LLM Integration (Local)
ollama>=0.3.0

# PDF Processing Libraries
PyPDF2>=3.0.0
pymupdf>=1.24.0
pdfplumber>=0.11.0

# LangChain for Text Processing
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# Image Processing
Pillow>=10.0.0

# Core Scientific Computing
numpy>=2.1.0
tqdm>=4.66.0

# HTTP and Utilities
requests>=2.32.0
httpx>=0.27.0
transformers>=4.45.0

# Configuration and Environment
python-dotenv>=1.0.0
pydantic>=2.5.0

# Optional: For advanced features
# nltk>=3.8.0                    # Natural language processing
# spacy>=3.7.0                   # Advanced NLP (requires model download)
# pandas>=2.1.0                  # Data analysis
# llama-cpp-python>=0.2.0        # Alternative to Ollama
# openai>=1.6.0                  # If using OpenAI API